requirements.txt
flask==2.0.1
openai==1.3.0
python-dotenv==1.0.0 